+++ src/GetVectorFromPoint.cpp
@@ -3,8 +3,8 @@
 
 void GetVectorFromPoint(sp& k, sp& l, int px, int py, const fsize* pView, int cx, int cy, matrix* const pMatrix)
 {
-	double rx = (pView->right - pView->left) * px / cx + pView->left;
-	double ry = (pView->bottom - pView->top) * py / cx + pView->top;
+	float rx = (pView->right - pView->left) * px / cx + pView->left;
+	float ry = (pView->bottom - pView->top) * py / cx + pView->top;
 
 	k = sp(0.01 * rx / (pView->right - pView->left), 0.01 * ry / (pView->bottom - pView->top), 0.01);
 	l = sp(rx, ry, -20);
