+++ src/Torus.h
@@ -11,8 +11,8 @@ class	Torus : public	Node
 protected:
 	DECLARE_SERIAL(Torus)
 private:
-	double	m_R;
-	double	m_r;
+	float	m_R;
+	float	m_r;
 public:
 	// コンストラクタ
 	Torus(Node* const root = 0, const char* const Name = "Torus", const sp Color = sp(255, 255, 255));
