+++ src/expand.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-expand::expand(	double r ) : matrix( 4, 4 )
+expand::expand(	float r ) : matrix( 4, 4 )
 {
 	set_data(1, 1, r);
 	set_data(2, 2, r);
