+++ src/info.h
@@ -8,11 +8,11 @@ struct Info
 {
 	D3DMATERIAL9	Material;
 	bool			isEnter;	// 入り込む
-	double			Distance;	// 交点までの距離
+	float			Distance;	// 交点までの距離
 	sp				Cross;		// 交点座標
 	sp				Vertical;	// 法線ベクトル
 	const Node *	pNode;		// 
-	double	Refractive;			// 屈折率
+	float	Refractive;			// 屈折率
 	Info()
 	{
 		pNode = NULL;
